adjust klinger volume oscillator
i don't think investopedia is right. the value seems to be opposite
of what data shows (and what the simpler logic typically shows).

+++ examples/file_json.rs
@@ -9,7 +9,7 @@ struct SecStats {
     low: Vec<f64>,
     open: Vec<f64>,
     close: Vec<f64>,
-    volume: Vec<f64>,
+    volume: Vec<u64>,
 }
 
 fn main() {
